# spec/api.yaml
openapi: 3.0.3
info:
  title: calpol
  description: Multi-Service health monitor
  contact:
    name: Jacob Halsey
    email: jacob@jhalsey.com
  license:
    name: ''
  version: 0.1.0
paths:
  /api/password_reset/:
    post:
      tags:
      - PasswordReset
      summary: Request a password reset token
      description: |
        Request a password reset token
      operationId: Request
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
        required: true
      responses:
        '200':
          description: Success
        default:
          $ref: '#/components/responses/ApiError'
      deprecated: false
      security: []
  /api/password_reset/submit:
    post:
      tags:
      - PasswordReset
      summary: Submit a password reset token
      description: |
        Submit a password reset token
      operationId: Submit
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitPasswordResetRequest'
        required: true
      responses:
        '200':
          description: Success
        default:
          $ref: '#/components/responses/ApiError'
      deprecated: false
      security: []
  /api/runner/queue:
    post:
      tags:
      - Runner
      summary: Queue the test runner to immediately re-run.
      description: |
        Queue the test runner to immediately re-run.
      operationId: Queue
      responses:
        '200':
          description: Success
        default:
          $ref: '#/components/responses/ApiError'
      deprecated: false
  /api/runner_logs:
    get:
      tags:
      - RunnerLogs
      summary: List the test runner logs
      description: |
        List the test runner logs
      operationId: ListRunnerLogs
      responses:
        '200':
          description: List of test runner logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListRunnerLogsResponse'
        default:
          $ref: '#/components/responses/ApiError'
      deprecated: false
  /api/sessions:
    get:
      tags:
      - Sessions
      summary: List this users active sessions
      description: |
        List this users active sessions
      operationId: ListSessions
      responses:
        '200':
          description: List of sessions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListSessionsResponse'
        default:
          $ref: '#/components/responses/ApiError'
      deprecated: false
  /api/sessions/login:
    post:
      tags:
      - Sessions
      summary: Login with email and password
      description: |
        Login with email and password
      operationId: Login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: User account and session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        default:
          $ref: '#/components/responses/ApiError'
      deprecated: false
      security: []
  /api/sessions/logout:
    delete:
      tags:
      - Sessions
      summary: Logout (delete) session
      description: |
        Logout (delete) session
      operationId: Logout
      responses:
        '200':
          description: Success
        default:
          $ref: '#/components/responses/ApiError'
      deprecated: false
  /api/sessions/{id}:
    delete:
      tags:
      - Sessions
      summary: Delete a session by id
      description: |
        Delete a session by id
      operationId: DeleteSession
      parameters:
      - name: id
        in: path
        description: Session id to delete
        required: true
        deprecated: false
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
        default:
          $ref: '#/components/responses/ApiError'
      deprecated: false
  /api/test_results:
    get:
      tags:
      - TestResults
      summary: List test results
      description: |
        List test results
      operationId: ListTestResults
      responses:
        '200':
          description: List of test results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTestResultsResponse'
        default:
          $ref: '#/components/responses/ApiError'
      deprecated: false
  /api/tests:
    get:
      tags:
      - Tests
      summary: List tests
      description: |
        List tests
      operationId: ListTests
      responses:
        '200':
          description: List of test results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTestsResponse'
        default:
          $ref: '#/components/responses/ApiError'
      deprecated: false
    post:
      tags:
      - Tests
      summary: Create a new test
      description: |
        Create a new test
      operationId: CreateTest
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTestRequest'
        required: true
      responses:
        '200':
          description: The created test
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestSummary'
        default:
          $ref: '#/components/responses/ApiError'
      deprecated: false
  /api/tests/{name}:
    get:
      tags:
      - Tests
      summary: Update a test
      description: |
        Update a test
      operationId: UpdateTest
      parameters:
      - name: name
        in: path
        description: Test name to update
        required: true
        deprecated: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTestRequest'
        required: true
      responses:
        '200':
          description: The created test
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestSummary'
        default:
          $ref: '#/components/responses/ApiError'
      deprecated: false
components:
  schemas:
    CreateTestRequest:
      type: object
      required:
      - name
      - config
      properties:
        config:
          type: object
        enabled:
          type: boolean
        failure_threshold:
          type: integer
          format: int32
        name:
          type: string
    ListRunnerLogsResponse:
      type: object
      required:
      - items
      - total
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RunnerLog'
        total:
          type: integer
          format: int64
    ListSessionsResponse:
      type: object
      required:
      - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SessionSummary'
    ListTestResultsResponse:
      type: object
      required:
      - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TestResultSummary'
    ListTestsResponse:
      type: object
      required:
      - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TestSummary'
    LoginRequest:
      type: object
      required:
      - email
      - password
      properties:
        email:
          type: string
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      required:
      - user
      - session
      - token
      properties:
        session:
          $ref: '#/components/schemas/SessionSummary'
        token:
          type: string
        user:
          $ref: '#/components/schemas/UserSummary'
    PasswordResetRequest:
      type: object
      required:
      - email
      properties:
        email:
          type: string
    RunnerLog:
      type: object
      required:
      - id
      - time_started
      - time_finished
      properties:
        failure_reason:
          type: string
        id:
          type: integer
          format: int32
        tests_failed:
          type: integer
          format: int32
        tests_passed:
          type: integer
          format: int32
        tests_skipped:
          type: integer
          format: int32
        time_finished:
          type: string
          format: date-time
        time_started:
          type: string
          format: date-time
    SessionSummary:
      type: object
      required:
      - id
      - created
      - last_used
      - last_ip
      - user_agent
      properties:
        created:
          type: integer
          format: int64
        id:
          type: integer
          format: int32
        last_ip:
          type: string
        last_used:
          type: integer
          format: int64
        user_agent:
          type: string
    SubmitPasswordResetRequest:
      type: object
      required:
      - token
      - new_password
      properties:
        new_password:
          type: string
          format: password
        token:
          type: string
    TestSummary:
      type: object
      required:
      - name
      - config
      - enabled
      - failure_threshold
      - failing
      properties:
        config:
          type: object
        enabled:
          type: boolean
        failing:
          type: boolean
        failure_threshold:
          type: integer
          format: int32
        name:
          type: string
    UpdateTestRequest:
      type: object
      required:
      - config
      properties:
        config:
          type: object
        enabled:
          type: boolean
        failure_threshold:
          type: integer
          format: int32
    UserSummary:
      type: object
      required:
      - id
      - name
      - email
      - sms_notifications
      - email_notifications
      properties:
        email:
          type: string
        email_notifications:
          type: boolean
        id:
          type: integer
          format: int32
        name:
          type: string
        phone_number:
          type: string
        sms_notifications:
          type: boolean
  responses:
    ApiError:
      description: An Api Problem (RFC7807)
      content:
        application/json:
          schema:
            type: object
            title: ApiError
            properties:
              detail:
                type: string
                description: |
                  A human readable explanation specific to this occurrence of the
                  problem.
                example: Connection to database timed out
              instance:
                type: string
                description: |-
                  An absolute URI that identifies the specific occurrence of the problem.
                  It may or may not yield further information if dereferenced.
              status:
                type: integer
                format: int32
                description: |
                  The HTTP status code generated by the origin server for this occurrence
                  of the problem.
                example: 503
              title:
                type: string
                description: |
                  A short, summary of the problem type. Written in english and readable
                  for engineers (usually not suited for non technical stakeholders and
                  not localized); example: Service Unavailable
              type:
                type: string
                description: |
                  An absolute URI that identifies the problem type.  When dereferenced,
                  it SHOULD provide human-readable documentation for the problem type
                  (e.g., using HTML).
  securitySchemes:
    basicAuth:
      type: http
      scheme: bearer
security:
- basicAuth: []